fix(backend): rename invalid convex analytics module
